Add very limited relative arc support
Problem:
The commands A and a are not supported and the parser just creates a sequence of moveTo commands

Analysis:
This real issue here is that iOS does not support the construction of arc in the same way SVG Paths need it to. An arc SVG Path consists of 7 parameters being:
- radius in x and y direction (to create ellipses)
- rotation relative to x-axis
- large-arc flag
- sweep flag
- absolute or relative end point coordinates
These information are much more detailed than the iOS implementation of addArc of UIBezierPath, which is only capable of drawing parts of a circle, as it only has the input parameters:
- center
- radius
- start and end angle
- clockwise/anticlockwise flag

Solution:
Add a very rudimentary support for relative arcs (command „a“). This can parse very specific SVG Paths in the following format:
„a rx ry 0 z 0 ex 0“
Where
- rx and ry are the radius of the circle and have to be the same value
- z determines if the top or bottom half of the circle are being added (0: top, 1: bottom)
- ex is the relative x position of the endPoint and has to be exactly double the radius (a.k.a. the diameter)
With this addition it is possible to parse half circles and add them to the returned bezier path

In order to parse a full circle the input has to be: „a rx ry 0 0 0 ex 0a rx ry 0 1 0 -ex 0“