And yet another quite specific adjustment
Problem:
Using the relative path closing (with "z") leaves the position at the last known position instead of the starting point (where the path being is closed to). However using the absolut path closing (with "Z") the position is reset to (0|0).

Analsysis:
There is no adjustment of the position on a path closing

Solution:
Instead of resetting the position to (0|0) or staying at the last position, the last move in the commands list is used as relative position for the close command. This puts the position to where the path is then actually being closed.